+++ sdv/__init__.py
@@ -11,9 +11,9 @@ __version__ = '0.1.0'
 import logging
 import os
 
-from sdv.DataNavigator import DataLoader, CSVDataLoader, DataNavigator
-from sdv.Modeler import Modeler
-from sdv.Sampler import Sampler
+from sdv.data_navigator import DataLoader, CSVDataLoader, DataNavigator
+from sdv.modeler import Modeler
+from sdv.sampler import Sampler
 
 ROOT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 FILE_SUFFIX = '.pkl'
